Moved existing getbooks function to a controller,also made delete and add controllers with connected web routes as well.
