+++ src/app/Http/Controllers/DatabaseUse/EditBookController.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Http\Controllers\DatabaseUse;
+
+use App\Http\Controllers\Controller;
+
+class EditBookController extends Controller
+{
+}
