+++ src/app/Http/Middleware/DuplicateCheck.php
@@ -1,0 +1,2 @@
+<?php
+//Middleware that checks if the proposed changes makes the book identical to another.
